Avoid recreating directories within session
The pmkdir function was designed to create subdirectories for any path.  Now
it's been redesigned to specifically create directories on the format XX/XX,
while remembering what directories it has already created in order to reduce
the number of redundant system calls.

The list of previously created directories is kept in bit arrays, so it only
occupies 8224 bytes of BSS space.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>