Fix behavior of ca-cas-fsck on large pack files
Instead of verifying the SHA-1 digest of the contained objects in hash order,
we sort up to 1 million objects at a time by their physical order, and test
them in that order.  This increased throughput from 30 MB/s to 115 MB/s on my
computer, and should also reduce mechanical wear on the hard disk drive.

Signed-off-by: Morten Hustveit <morten.hustveit@gmail.com>